// Assets-Uncapped/front-portal-develop/src/services/api/partners/models/CompanyLegalDetailsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-partners
 * Change log: <br>1.0.0 initial release
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CompanyAddress } from './CompanyAddress';
import {
    CompanyAddressFromJSON,
    CompanyAddressFromJSONTyped,
    CompanyAddressToJSON,
    CompanyAddressToJSONTyped,
} from './CompanyAddress';

/**
 * 
 * @export
 * @interface CompanyLegalDetailsResponse
 */
export interface CompanyLegalDetailsResponse {
    /**
     * 
     * @type {string}
     * @memberof CompanyLegalDetailsResponse
     */
    companyName?: string;
    /**
     * 
     * @type {string}
     * @memberof CompanyLegalDetailsResponse
     */
    type?: string;
    /**
     * 
     * @type {Date}
     * @memberof CompanyLegalDetailsResponse
     */
    dateOfCreation?: Date;
    /**
     * 
     * @type {string}
     * @memberof CompanyLegalDetailsResponse
     */
    companyNumber?: string;
    /**
     * 
     * @type {CompanyAddress}
     * @memberof CompanyLegalDetailsResponse
     */
    registeredAddress?: CompanyAddress;
    /**
     * 
     * @type {CompanyAddress}
     * @memberof CompanyLegalDetailsResponse
     */
    billingAddress?: CompanyAddress;
    /**
     * 
     * @type {boolean}
     * @memberof CompanyLegalDetailsResponse
     */
    sameAddresses?: boolean;
}

/**
 * Check if a given object implements the CompanyLegalDetailsResponse interface.
 */
export function instanceOfCompanyLegalDetailsResponse(value: object): value is CompanyLegalDetailsResponse {
    return true;
}

export function CompanyLegalDetailsResponseFromJSON(json: any): CompanyLegalDetailsResponse {
    return CompanyLegalDetailsResponseFromJSONTyped(json, false);
}

export function CompanyLegalDetailsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CompanyLegalDetailsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'companyName': json['companyName'] == null ? undefined : json['companyName'],
        'type': json['type'] == null ? undefined : json['type'],
        'dateOfCreation': json['dateOfCreation'] == null ? undefined : (new Date(json['dateOfCreation'])),
        'companyNumber': json['companyNumber'] == null ? undefined : json['companyNumber'],
        'registeredAddress': json['registeredAddress'] == null ? undefined : CompanyAddressFromJSON(json['registeredAddress']),
        'billingAddress': json['billingAddress'] == null ? undefined : CompanyAddressFromJSON(json['billingAddress']),
        'sameAddresses': json['sameAddresses'] == null ? undefined : json['sameAddresses'],
    };
}

export function CompanyLegalDetailsResponseToJSON(json: any): CompanyLegalDetailsResponse {
    return CompanyLegalDetailsResponseToJSONTyped(json, false);
}

export function CompanyLegalDetailsResponseToJSONTyped(value?: CompanyLegalDetailsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'companyName': value['companyName'],
        'type': value['type'],
        'dateOfCreation': value['dateOfCreation'] == null ? undefined : ((value['dateOfCreation']).toISOString().substring(0,10)),
        'companyNumber': value['companyNumber'],
        'registeredAddress': CompanyAddressToJSON(value['registeredAddress']),
        'billingAddress': CompanyAddressToJSON(value['billingAddress']),
        'sameAddresses': value['sameAddresses'],
    };
}